TeachBase Lesson_02 Task_06

+++ teach_base/lesson_02/06_sum_of_purchases.rb
@@ -21,7 +21,7 @@ loop do
   puts "Enter price of #{prod}!"
   price = gets.chomp.to_f
   puts 'Enter how match product you want to buy!'
-  count = gets.chomp.to_f
+  count = gets.chomp.to_i
   basket[prod] = {
     price: price,
     count: count,
